+++ src/Config.php
@@ -4,13 +4,13 @@ namespace ABSmartly\SDK;
 
 use ABSmartly\SDK\Client\Client;
 use ABSmartly\SDK\Context\ContextDataProvider;
-use ABSmartly\SDK\Context\ContextEventHandler;
+use ABSmartly\SDK\Context\ContextPublisher;
 use ABSmartly\SDK\Context\ContextEventLogger;
 
 class Config {
 	private Client $client;
 	private ContextDataProvider $contextDataProvider;
-	private ContextEventHandler $contextEventHandler;
+	private ContextPublisher $contextEventHandler;
 	private ?ContextEventLogger $contextEventLogger = null;
 
 	public function __construct(Client $client) {
@@ -26,7 +26,7 @@ class Config {
 		return $this;
 	}
 
-	public function setContextEventHandler(ContextEventHandler $contextEventHandler): Config {
+	public function setContextEventHandler(ContextPublisher $contextEventHandler): Config {
 		$this->contextEventHandler = $contextEventHandler;
 		return $this;
 	}
@@ -38,9 +38,9 @@ class Config {
 		return $this->contextDataProvider;
 	}
 
-	public function getContextEventHandler(): ContextEventHandler {
+	public function getContextEventHandler(): ContextPublisher {
 		if (!isset($this->contextEventHandler)) {
-			$this->contextEventHandler = new ContextEventHandler($this->client);
+			$this->contextEventHandler = new ContextPublisher($this->client);
 		}
 
 		return $this->contextEventHandler;
