+++ src/Context/ContextEventHandler.php
@@ -2,17 +2,8 @@
 
 namespace ABSmartly\SDK\Context;
 
-use ABSmartly\SDK\Client\ClientInterface;
-use ABSmartly\SDK\PublishEvent;
-
-class ContextEventHandler {
-	private ClientInterface $client;
-
-	public function __construct(ClientInterface $client) {
-		$this->client = $client;
-	}
-
-	public function publish(PublishEvent $event): void {
-		$this->client->publish($event);
-	}
+/**
+ * @deprecated Use ContextPublisher instead.
+ */
+class ContextEventHandler extends ContextPublisher {
 }
